requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
alembic==1.7.7
    # via flask-migrate
async-timeout==4.0.2
    # via redis
bidict==0.22.0
    # via python-socketio
cachecontrol==0.12.10
    # via firebase-admin
cachelib==0.10.2
    # via flask-session
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via flask
colorama==0.4.4
    # via
    #   click
    #   loguru
coverage==6.4.3
    # via -r requirements.in
deprecated==1.2.13
    # via redis
dnspython==2.2.1
    # via eventlet
eventlet==0.33.0
    # via -r requirements.in
firebase-admin==5.2.0
    # via -r requirements.in
flake8==4.0.1
    # via -r requirements.in
flask==2.1.1
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-marshmallow
    #   flask-migrate
    #   flask-session
    #   flask-socketio
    #   flask-sqlalchemy
flask-cors==3.0.10
    # via -r requirements.in
flask-marshmallow==0.14.0
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-session==0.4.1
    # via -r requirements.in
flask-socketio==5.3.1
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
geoalchemy2==0.11.1
    # via -r requirements.in
google-api-core[grpc]==2.7.2
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.44.0
    # via firebase-admin
google-auth==2.6.5
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.0
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.4.0
    # via firebase-admin
google-cloud-storage==2.3.0
    # via firebase-admin
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-storage
googleapis-common-protos==1.56.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==1.1.2
    # via
    #   eventlet
    #   sqlalchemy
grpcio==1.44.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.44.0
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
huey==2.4.3
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via
    #   alembic
    #   flask
importlib-resources==5.7.1
    # via alembic
itsdangerous==2.1.2
    # via flask
jinja2==3.1.1
    # via flask
loguru==0.6.0
    # via -r requirements.in
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
marshmallow==3.15.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.27.0
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
msgpack==1.0.3
    # via cachecontrol
packaging==21.3
    # via
    #   geoalchemy2
    #   marshmallow
    #   redis
proto-plus==1.20.3
    # via google-cloud-firestore
protobuf==3.20.0
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2==2.9.3
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pygeohash==1.2.0
    # via -r requirements.in
pyparsing==3.0.8
    # via
    #   httplib2
    #   packaging
python-dotenv==0.19.2
    # via -r requirements.in
python-engineio==4.3.1
    # via python-socketio
python-socketio==5.5.2
    # via flask-socketio
redis==4.2.2
    # via
    #   -r requirements.in
    #   walrus
requests==2.27.1
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   eventlet
    #   flask-cors
    #   flask-marshmallow
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
sqlalchemy==1.4.35
    # via
    #   alembic
    #   flask-sqlalchemy
    #   geoalchemy2
    #   marshmallow-sqlalchemy
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.9
    # via requests
walrus==0.9.2
    # via -r requirements.in
werkzeug==2.1.1
    # via flask
win32-setctime==1.1.0
    # via loguru
wrapt==1.14.0
    # via deprecated
zipp==3.8.0
    # via
    #   importlib-metadata
    #   importlib-resources
